Move backend/sftp.StartForeground to backend/
